+++ src/main/java/use_case/edit_review/EditOutputData.java
@@ -3,12 +3,12 @@ package use_case.edit_review;
 public class EditOutputData {
 
     private final String comment;
-    private final int rating;
+    private final double rating;
     private final String username;
     private final int songId;
     private final String songName;
 
-    public EditOutputData(String comment, int rating, String username, int songId, String songName) {
+    public EditOutputData(String comment, double rating, String username, int songId, String songName) {
         this.comment = comment;
         this.rating = rating;
         this.username = username;
@@ -20,7 +20,7 @@ public class EditOutputData {
         return comment;
     }
 
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
